Extract plugin loading and marketplace logic into codex-core-plugins
