+++ codex-rs/core-plugins/src/lib.rs
@@ -1,0 +1,49 @@
+pub mod loader;
+pub mod manifest;
+pub mod marketplace;
+pub mod remote;
+pub mod store;
+pub mod toggles;
+
+pub use loader::ResolvedPluginSkills;
+pub use loader::configured_curated_plugin_ids_from_codex_home;
+pub use loader::installed_plugin_telemetry_metadata;
+pub use loader::load_plugin_apps;
+pub use loader::load_plugin_mcp_servers;
+pub use loader::load_plugin_skills;
+pub use loader::load_plugins_from_layer_stack;
+pub use loader::log_plugin_load_errors;
+pub use loader::plugin_telemetry_metadata_from_root;
+pub use loader::refresh_curated_plugin_cache;
+pub use loader::refresh_non_curated_plugin_cache;
+pub use manifest::PluginManifestInterface;
+pub use manifest::PluginManifestPaths;
+pub use manifest::load_plugin_manifest;
+pub use marketplace::Marketplace;
+pub use marketplace::MarketplaceError;
+pub use marketplace::MarketplaceInterface;
+pub use marketplace::MarketplaceListError;
+pub use marketplace::MarketplaceListOutcome;
+pub use marketplace::MarketplacePlugin;
+pub use marketplace::MarketplacePluginAuthPolicy;
+pub use marketplace::MarketplacePluginInstallPolicy;
+pub use marketplace::MarketplacePluginPolicy;
+pub use marketplace::MarketplacePluginSource;
+pub use marketplace::ResolvedMarketplacePlugin;
+pub use marketplace::list_marketplaces;
+pub use marketplace::load_marketplace;
+pub use marketplace::resolve_marketplace_plugin;
+pub use marketplace::validate_marketplace_root;
+pub use remote::RemotePluginFetchError;
+pub use remote::RemotePluginMutationError;
+pub use remote::RemotePluginServiceConfig;
+pub use remote::RemotePluginStatusSummary;
+pub use remote::enable_remote_plugin;
+pub use remote::fetch_remote_featured_plugin_ids;
+pub use remote::fetch_remote_plugin_status;
+pub use remote::uninstall_remote_plugin;
+pub use store::PluginInstallResult;
+pub use store::PluginStore;
+pub use store::PluginStoreError;
+pub use store::plugin_version_for_source;
+pub use toggles::collect_plugin_enabled_candidates;
