+++ codex-rs/core/src/plugins/mod.rs
@@ -16,6 +16,7 @@ mod store;
 pub(crate) mod test_support;
 mod toggles;
 
+pub use codex_core_plugins::RemotePluginServiceConfig;
 pub use codex_plugin::AppConnectorId;
 pub use codex_plugin::EffectiveSkillRoots;
 pub use codex_plugin::PluginCapabilitySummary;
@@ -27,6 +28,16 @@ pub use codex_plugin::validate_plugin_segment;
 pub type LoadedPlugin = codex_plugin::LoadedPlugin<McpServerConfig>;
 pub type PluginLoadOutcome = codex_plugin::PluginLoadOutcome<McpServerConfig>;
 
+pub(crate) use codex_core_plugins::configured_curated_plugin_ids_from_codex_home;
+pub use codex_core_plugins::installed_plugin_telemetry_metadata;
+pub use codex_core_plugins::load_plugin_apps;
+pub use codex_core_plugins::load_plugin_mcp_servers;
+pub(crate) use codex_core_plugins::load_plugin_skills;
+pub(crate) use codex_core_plugins::load_plugins_from_layer_stack;
+pub(crate) use codex_core_plugins::log_plugin_load_errors;
+pub use codex_core_plugins::plugin_telemetry_metadata_from_root;
+pub(crate) use codex_core_plugins::refresh_curated_plugin_cache;
+pub(crate) use codex_core_plugins::refresh_non_curated_plugin_cache;
 pub(crate) use discoverable::list_tool_suggest_discoverable_plugins;
 pub(crate) use injection::build_plugin_injections;
 pub use installed_marketplaces::INSTALLED_MARKETPLACES_DIR;
@@ -45,12 +56,7 @@ pub use manager::PluginRemoteSyncError;
 pub use manager::PluginUninstallError;
 pub use manager::PluginsManager;
 pub use manager::RemotePluginSyncResult;
-pub use manager::installed_plugin_telemetry_metadata;
-pub use manager::load_plugin_apps;
-pub use manager::load_plugin_mcp_servers;
-pub use manager::plugin_telemetry_metadata_from_root;
 pub use manifest::PluginManifestInterface;
-pub(crate) use manifest::PluginManifestPaths;
 pub(crate) use manifest::load_plugin_manifest;
 pub use marketplace::MarketplaceError;
 pub use marketplace::MarketplaceListError;
